Filtering database search results by field
New 'args' being passed through the searching invoke. Works for strings/integers/etc. but is finnicky for more complicated fields. To be done: reserve the first argument of args for something that specifies the table to be searched. Note that this solution requires data table to be completely queried and rebuilt instead of sorting a local copy, but current is probably the prefered method.